+++ packages/dso-toolkit/src/components/modal/modal.stories-of.ts
@@ -5,6 +5,7 @@ import { Modal } from "./modal.models.js";
 
 // <see #1550>
 let onLeave: { load: () => void; unload: () => void } | null = null;
+
 function onStory(load: () => void, unload: () => void) {
   onLeave = { load, unload };
 
@@ -14,6 +15,40 @@ function onStory(load: () => void, unload: () => void) {
 const storyObserver = new MutationObserver(([titleMutationRecord]) => {
   if (titleMutationRecord.target.textContent?.startsWith("HTML|CSS / Modal")) {
     toggleClass("dso-modal-open");
+  }
+
+  if (titleMutationRecord.target.textContent?.startsWith("Core / Modal")) {
+    setTimeout(
+      () =>
+        onStory(
+          () => null,
+          // eslint-disable-next-line @typescript-eslint/no-explicit-any
+          () => (window as any).modalRef.close()
+        ),
+      400
+    );
+  }
+
+  if (titleMutationRecord.target.textContent?.startsWith("Modal")) {
+    setTimeout(
+      () =>
+        onStory(
+          () => null,
+          () => {
+            // eslint-disable-next-line @typescript-eslint/no-explicit-any
+            const modalRef = (window as any).modalRef;
+
+            if (modalRef.framework === "angular") {
+              modalRef.ref.close();
+            }
+
+            if (modalRef.framework === "react") {
+              modalRef.ref.remove();
+            }
+          }
+        ),
+      400
+    );
   }
 
   if (onLeave) {
@@ -55,7 +90,6 @@ function toggleClass(className: string) {
 
 export interface ModalTemplates<TemplateFnReturnType> {
   modalTemplate: (modalProperties: Modal<TemplateFnReturnType>) => TemplateFnReturnType;
-  modalControllerTemplate: (modalProperties: Modal<TemplateFnReturnType>) => TemplateFnReturnType;
   activeBody: TemplateFnReturnType;
   activeFooter: TemplateFnReturnType;
   passiveBody: TemplateFnReturnType;
@@ -78,19 +112,6 @@ export function storiesOfModal<Implementation, Templates, TemplateFnReturnType>(
     stories.addParameters({
       argTypes: modalArgTypes,
     });
-
-    stories.add(
-      "controller",
-      templateMapper<ModalArgs>((args, { modalControllerTemplate, confirmBody, confirmFooter }) =>
-        modalControllerTemplate(modalArgsMapper(args, confirmBody, confirmFooter))
-      ),
-      {
-        args: componentArgs<Pick<ModalArgs, "role" | "modalTitle">>({
-          role: "dialog",
-          modalTitle: "Disclaimer",
-        }),
-      }
-    );
 
     stories.add(
       "confirm",